+++ src/main/java/subway/Exception/SectionExceptionHandler.java
@@ -1,6 +1,6 @@
 package subway.Exception;
 
-import subway.Controller.LineController;
+import subway.Controller.SectionController;
 import subway.domain.LineRepository;
 import subway.view.OutputView;
 
@@ -9,12 +9,58 @@ public class SectionExceptionHandler {
     private static final int MIN_LENGTH = 2;
 
     private static void goBack() {
-        LineController.run();
+        SectionController.run();
     }
 
     private static void printErrorAndGoBack(String error) {
         OutputView.printError(error);
         goBack();
+    }
+
+    public static void unselectable(String selection, String[] pattern) {
+        try {
+            ExceptionHandler.isUnSelectable(selection, pattern);
+        } catch (CustomException e) {
+
+        }
+    }
+
+    public static void positiveInt(String string) {
+        try {
+            isNotInt(string);
+            isNotPositive(string);
+        } catch (CustomException e) {
+            OutputView.printError(e.getMessage());
+            SectionController.run();
+        }
+    }
+
+    private static void isNotInt(String string) {
+        try {
+            Integer.parseInt(string);
+        } catch (Exception e) {
+            throw new CustomException("숫자형식이 아닙니다.");
+        }
+    }
+
+    private static void isNotPositive(String string) {
+        if (Integer.parseInt(string) <= 0) {
+            throw new CustomException("0보다 큰 수이어야 합니다.");
+        }
+    }
+
+    public static void notLineContained(String lineName) {
+        try {
+            isNotLineContained(lineName);
+        } catch (CustomException e) {
+            printErrorAndGoBack(e.getMessage());
+        }
+    }
+
+    private static void isNotLineContained(String lineName) {
+        if (!LineRepository.contains(lineName)) {
+            throw new CustomException("등록되지 않은 노선에 구간을 추가할 수 없습니다.");
+        }
     }
 
     public static void noLine(String lineName) {
@@ -45,6 +91,38 @@ public class SectionExceptionHandler {
             }
         } catch (CustomException e) {
             printErrorAndGoBack(e.getMessage());
+        }
+    }
+
+    public static void stationShouldNotBeInLineForAddingToSection(String lineName,
+        String stationName) {
+        try {
+            isStationInLine(lineName, stationName);
+        } catch (CustomException e) {
+            OutputView.printError(e.getMessage());
+            SectionController.run();
+        }
+    }
+
+    private static void isStationInLine(String lineName, String stationName) {
+        if (LineRepository.lineContainsStation(lineName, stationName)) {
+            throw new CustomException("이미 구간에 있는 역은 추가할 수 없습니다.");
+        }
+    }
+
+    public static void lineOverflow(String line, String order) {
+        try {
+            overflow(line, Integer.parseInt(order));
+        } catch (CustomException e) {
+            OutputView.printError(e.getMessage());
+            SectionController.run();
+        }
+    }
+
+    private static void overflow(String line, int order) {
+        int max = LineRepository.getSectionLength(line) - 1;
+        if (order > max) {
+            throw new CustomException(max + "보다 작아야 합니다.");
         }
     }
 }
